+++ src/chart_plugin/ui_helpers/components.rs
@@ -29,6 +29,11 @@ pub struct SelectedTab {
     pub id: ReflectableUuid,
 }
 
+#[derive(Component, Default)]
+pub struct ScrollingList {
+    pub position: f32,
+}
+
 #[derive(Component)]
 pub struct SelectedTabTextInput {
     pub id: ReflectableUuid,
